+++ CHANGELOG.md
@@ -2,10 +2,16 @@
 
 ## Unreleased
 
+#### Changes
+
+- [BREAKING] Removed serialization of AST structs (#1442).
+
 #### Enhancements
 
-- Exposed some pretty printing internals for `MastNode`
-- Made `KernelLibrary` impl `Clone` and `AsRef<Library>`
+- Exposed some pretty printing internals for `MastNode` (#1441).
+- Made `KernelLibrary` impl `Clone` and `AsRef<Library>` (#1441).
+- Added serialization to the `Program` struct (#1442).
+
 
 ## 0.10.0 (2024-08-06)